important fixes

+++ query.h
@@ -42,7 +42,7 @@ public:
 
     void printAnswers(){
         cout << "We found these files for you: " << endl;
-        for(int i = 0; i<answers.size(); i++){
+        for(size_t i = 0; i<answers.size(); i++){
             cout << answers[i] << ", ";
             if(i%5 == 0){
                 cout << endl;
